wikum/requirements.txt
aioredis==1.3.0
amqp==2.5.1
appnope==0.1.0
asgiref==2.3.2
asn1crypto==1.0.1
astroid==2.3.1
async-timeout==3.0.1
attrs==19.2.0
autobahn==19.10.1
Automat==0.7.0
backcall==0.1.0
backports.functools-lru-cache==1.5
billiard==3.6.1.0
bleach==2.1.4
boto==2.49.0
boto3==1.9.246
botocore==1.12.246
breadability==0.1.20
bz2file==0.98
caniusepython3==7.1.0
celery==4.3.0
certifi==2019.9.11
cffi==1.12.3
channels==2.1.3
channels-redis==2.3.0
chardet==3.0.4
constantly==15.1.0
coverage==4.5.4
cryptography==2.7
cycler==0.10.0
daphne==2.3.0
decorator==4.4.0
distlib==0.2.9.post0
Django==2.1.1
django-annoying==0.10.5
django-appconf==1.0.3
django-tracking2==0.5.0
django-user-accounts==2.1.0
docopt==0.6.2
docutils==0.15.2
future==0.18.0
gnureadline==8.0.0
hiredis==1.0.0
html5lib==1.0.1
httpretty==0.9.7
hyperlink==19.0.0
idna==2.8
importlib-metadata==0.23
incremental==17.5.0
ipython==7.8.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.15.1
jmespath==0.9.4
joblib==0.14.0
jsonschema==3.1.1
kombu==4.6.5
lazy-object-proxy==1.4.2
lxml==4.4.1
mccabe==0.6.1
more-itertools==7.2.0
msgpack==0.5.6
mwparserfromhell==0.5.4
mysql==0.0.2
mysqlclient==1.4.4
nltk==3.4.5
numpy==1.17.2
packaging==19.2
pandas==0.25.1
parso==0.5.1
path.py==12.0.1
pbr==5.4.3
pexpect==4.7.0
pickleshare==0.7.5
praw==6.4.0
prawcore==1.0.1
prompt-toolkit==2.0.10
ptyprocess==0.6.0
py-wikimarkup==2.0.0
pycountry==19.8.18
pycparser==2.19
Pygments==2.4.2
PyHamcrest==1.9.0
pylint==2.4.2
pyparsing==2.4.2
pyrsistent==0.15.4
python-dateutil==2.8.0
python-dotenv==0.10.3
pytz==2019.3
requests==2.22.0
s3transfer==0.2.1
scikit-learn==0.21.3
scipy==1.3.1
simplegeneric==0.8.1
six==1.12.0
smart-open==1.8.4
SQLAlchemy==1.3.10
sumy==0.8.1
traitlets==4.3.3
Twisted==19.7.0
txaio==18.8.1
typed-ast==1.4.0
Unidecode==1.1.1
update-checker==0.16
urllib3==1.25.6
vine==1.3.0
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.56.0
wikitools==1.2
wrapt==1.11.2
zipp==0.6.0
zope.interface==4.6.0
